backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0.post1
sqlalchemy==2.0.23
alembic==1.13.0
psycopg2-binary==2.9.9
pydantic==2.5.2
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
celery==5.3.6
redis==5.0.1
docker==6.1.3
nbformat==5.9.2
nbconvert==7.11.0
jupyter-client==8.6.0
ipykernel==6.26.0
emails==0.6
httpx==0.25.2
pytest==7.4.3
